requirements.txt
pandas>=1.3.3
requests>=2.26.0
flake8>=3.9.2
bandit>=1.7.0
virtualenv>=20.24.2
python-dotenv>=0.19.1
pytest>=6.2.5
jsonschema>=4.18.4
splunk-sdk>=1.7.4
github>=3.4.0
